requirement.txt
asgiref==3.5.2
astroid==2.12.2
atomicwrites==1.4.1
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Django==4.0.6
dominate==2.6.0
Flask==2.1.3
Flask-Bootstrap4==4.0.2
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
mysql-connector-python==8.0.29
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
protobuf==4.21.2
py==1.11.0
PyLin==0.2.7
pylint==2.14.4
pyparsing==3.0.9
pyserial==3.5
requests==2.28.1
SQLAlchemy==1.4.39
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.10
visitor==0.1.3
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.1
